requirements.txt
streamlit
requests
flask
flask-cors
mysql-connector-python
python-dotenv